// intranet/src/app/core/helios-api-client/model/centreModule.ts
/**
 * Helios API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CentreInfos } from './centreInfos';
import { Module } from './module';


export interface CentreModule { 
    centre: CentreInfos;
    modules: Array<Module> | null;
}