// intranet/src/app/core/helios-api-client/model/module.ts
/**
 * Helios API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Modules } from './modules';


export interface Module { 
    label: string;
    path: string;
    parentId?: number | null;
    sousMenus?: Array<any> | null;
    icon?: string | null;
    title?: string | null;
    prefixIcon?: string | null;
    suffixIcon?: string | null;
    id?: number;
    description?: string | null;
    code?: Modules;
    creation?: string;
    modification?: string | null;
}
export namespace Module {
}